refactor: documents.py to blender.py, move functions
