+++ src/nitpick/generic.py
@@ -5,15 +5,11 @@
     from nitpick.generic import *
 """
 import re
-from collections import OrderedDict
 from pathlib import Path
-from typing import Any, Dict, Iterable, List, MutableMapping, Optional, Tuple, Union
+from typing import Iterable, List, Optional, Tuple
 
-import jmespath
-from jmespath.parser import ParsedResult
-
-from nitpick.constants import DOUBLE_QUOTE, PROJECT_NAME, SEPARATOR_FLATTEN, SEPARATOR_QUOTED_SPLIT, SINGLE_QUOTE
-from nitpick.typedefs import JsonDict, PathOrStr
+from nitpick.constants import DOUBLE_QUOTE, PROJECT_NAME, SEPARATOR_QUOTED_SPLIT, SINGLE_QUOTE
+from nitpick.typedefs import PathOrStr
 
 URL_RE = re.compile(r"[a-z]+://[\$\w]\w*")
 
@@ -25,35 +21,6 @@ def get_subclasses(cls):
         subclasses.append(subclass)
         subclasses += get_subclasses(subclass)
     return subclasses
-
-
-# TODO: move flatten() and unflatten() to DictBlender: they depend on each other and keep state between calls.
-def flatten(dict_, parent_key="", separator=".", current_lists=None) -> JsonDict:
-    """Flatten a nested dict.
-
-    Use :py:meth:`unflatten()` to revert.
-
-    Adapted from `this StackOverflow question <https://stackoverflow.com/a/6027615>`_.
-    """
-    if current_lists is None:
-        current_lists = {}
-
-    items: List[Tuple[str, Any]] = []
-    for key, value in dict_.items():
-        quoted_key = f"{DOUBLE_QUOTE}{key}{DOUBLE_QUOTE}" if separator in str(key) else key
-        new_key = str(parent_key) + separator + str(quoted_key) if parent_key else quoted_key
-        if isinstance(value, dict):
-            items.extend(flatten(value, new_key, separator, current_lists).items())
-        elif isinstance(value, (list, tuple)):
-            # If the value is a list or tuple, append to a previously existing list.
-            existing_list = current_lists.get(new_key, [])
-            existing_list.extend(list(value))
-            current_lists[new_key] = existing_list
-
-            items.append((new_key, existing_list))
-        else:
-            items.append((new_key, value))
-    return dict(items)
 
 
 def quoted_split(string_: str, separator=".") -> List[str]:
@@ -88,113 +55,6 @@ def quoted_split(string_: str, separator=".") -> List[str]:
         part.replace(SEPARATOR_QUOTED_SPLIT, separator)
         for part in quoted_regex.sub(remove_quotes, string_).split(separator)
     ]
-
-
-def unflatten(dict_, separator=".", sort=True) -> OrderedDict:
-    """Turn back a flattened dict created by :py:meth:`flatten()` into a nested dict.
-
-    >>> expected = {'my': {'sub': {'path': True}, 'home': 4}, 'another': {'path': 3}}
-    >>> unflatten({"my.sub.path": True, "another.path": 3, "my.home": 4}) == expected
-    True
-    >>> unflatten({"repo": "conflicted key", "repo.name": "?", "repo.path": "?"})
-    Traceback (most recent call last):
-      ...
-    TypeError: 'str' object does not support item assignment
-    """
-    items: OrderedDict = OrderedDict()
-    for root_key, root_value in sorted(dict_.items()) if sort else dict_.items():
-        all_keys = quoted_split(root_key, separator)
-        sub_items = items
-        for key in all_keys[:-1]:
-            try:
-                sub_items = sub_items[key]
-            except KeyError:
-                sub_items[key] = OrderedDict()
-                sub_items = sub_items[key]
-
-        sub_items[all_keys[-1]] = root_value
-
-    return items
-
-
-class DictBlender:
-    """A blender of dictionaries: keep adding dictionaries and mix them all at the end.
-
-    .. note::
-
-        This class intentionally doesn't inherit from the standard ``dict()``.
-        It's an unnecessary hassle to override and deal with all those magic dunder methods.
-    """
-
-    def __init__(self, original_dict: JsonDict = None) -> None:
-        self._all_flattened: OrderedDict = OrderedDict()
-        self._current_lists: Dict[str, Iterable] = {}
-        self.add(original_dict or {})
-
-    def add(self, other: JsonDict) -> None:
-        """Add another dictionary to the existing data."""
-        flattened_other = flatten(other, separator=SEPARATOR_FLATTEN, current_lists=self._current_lists)
-        self._all_flattened.update(flattened_other)
-
-    def mix(self, sort=True) -> JsonDict:
-        """Mix all dictionaries, replacing values with identical keys and extending lists."""
-        return unflatten(self._all_flattened, separator=SEPARATOR_FLATTEN, sort=sort)
-
-
-def find_object_by_key(list_: List[dict], search_key: str, search_value: Any) -> dict:
-    """Find an object in a list, using a key/value pair to search.
-
-    >>> fruits = [{"id": 1, "fruit": "banana"}, {"id": 2, "fruit": "apple"}, {"id": 3, "fruit": "mango"}]
-    >>> find_object_by_key(fruits, "id", 1) == {'id': 1, 'fruit': 'banana'}
-    True
-    >>> find_object_by_key(fruits, "fruit", "banana") == {'id': 1, 'fruit': 'banana'}
-    True
-    >>> find_object_by_key(fruits, "fruit", "pear")
-    {}
-    >>> find_object_by_key(fruits, "fruit", "mango") == {'id': 3, 'fruit': 'mango'}
-    True
-    >>> find_object_by_key(None, "fruit", "pear")
-    {}
-    """
-    for obj in list_ or []:
-        if obj.get(search_key) == search_value:
-            return obj
-    return {}
-
-
-def jmes_search_json(
-    json_data: Union[MutableMapping[str, Any], List[Any]],
-    jmespath_expression: Union[ParsedResult, str],
-    default: Any = None,
-) -> Any:
-    """Search a dictionary or list using a JMESPath expression. Return a default value if not found.
-
-    >>> data = {"root": {"app": [1, 2], "test": "something"}}
-    >>> jmes_search_json(data, "root.app", None)
-    [1, 2]
-    >>> jmes_search_json(data, "root.test", None)
-    'something'
-    >>> jmes_search_json(data, "root.unknown", "")
-    ''
-    >>> jmes_search_json(data, "root.unknown", None)
-
-    >>> jmes_search_json(data, "root.unknown")
-
-    >>> jmes_search_json(data, jmespath.compile("root.app"), [])
-    [1, 2]
-    >>> jmes_search_json(data, jmespath.compile("root.whatever"), "xxx")
-    'xxx'
-
-    :param jmespath_expression: A compiled JMESPath expression or a string with an expression.
-    :param json_data: The dictionary to be searched.
-    :param default: Default value in case nothing is found.
-    :return: The object that was found or the default value.
-    """
-    if isinstance(jmespath_expression, str):
-        rv = jmespath.search(jmespath_expression, json_data)
-    else:
-        rv = jmespath_expression.search(json_data)
-    return rv or default
 
 
 def version_to_tuple(version: str = None) -> Tuple[int, ...]:
