+++ src/nitpick/project.py
@@ -17,6 +17,7 @@ from tomlkit.items import KeyType, SingleKey
 from tomlkit.toml_document import TOMLDocument
 
 from nitpick import fields, plugins
+from nitpick.blender import TomlDoc, search_json
 from nitpick.constants import (
     CONFIG_FILES,
     DOT_NITPICK_TOML,
@@ -30,9 +31,8 @@ from nitpick.constants import (
     TOOL_NITPICK_JMEX,
     TOOL_NITPICK_KEY,
 )
-from nitpick.documents import TomlDoc
 from nitpick.exceptions import QuitComplainingError
-from nitpick.generic import jmes_search_json, version_to_tuple
+from nitpick.generic import version_to_tuple
 from nitpick.schemas import BaseNitpickSchema, flatten_marshmallow_errors, help_message
 from nitpick.typedefs import JsonDict, PathOrStr, mypy_property
 from nitpick.violations import Fuss, ProjectViolations, Reporter, StyleViolations
@@ -149,7 +149,7 @@ class Project:
             return Configuration(None, [], "")
 
         toml_doc = TomlDoc(path=config_file)
-        config_dict = jmes_search_json(toml_doc.as_object, TOOL_NITPICK_JMEX, {})
+        config_dict = search_json(toml_doc.as_object, TOOL_NITPICK_JMEX, {})
         validation_errors = ToolNitpickSectionSchema().validate(config_dict)
         if not validation_errors:
             return Configuration(config_file, config_dict.get("style", []), config_dict.get("cache", ""))
@@ -181,7 +181,7 @@ class Project:
 
         from nitpick.flake8 import NitpickFlake8Extension
 
-        minimum_version = jmes_search_json(self.style_dict, NITPICK_MINIMUM_VERSION_JMEX, None)
+        minimum_version = search_json(self.style_dict, NITPICK_MINIMUM_VERSION_JMEX, None)
         logger.debug(f"Minimum version: {minimum_version}")
         if minimum_version and version_to_tuple(NitpickFlake8Extension.version) < version_to_tuple(minimum_version):
             yield Reporter().make_fuss(
